pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "runze_pump"
description = "a python driver for Aurora Pro Scientific or Runze Fluid SY-01B syringe pumps"
dynamic = ["version"]
license = {text= "MIT"}
authors = [
    {name = "Joshua Vasquez", email = "joshua.vasquez@alleninstitute.org"},
]
readme = "README.md"
keywords = ['fluidics', 'Runze Fluid', 'Aurora Pro Scientfic']
classifiers=[
    'License :: OSI Approved :: MIT License',
    'Development Status :: 4 - Beta',
    'Programming Language :: Python'
]

requires-python = ">=3.8"
dependencies = [
    "pyserial",
]

[project.optional-dependencies]
dev = [
    "sphinx",
    "furo",
    "enum-tools[sphinx]",
]

[project.urls]
repository = "https://github.com/AllenNeuralDynamics/runze-control"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "runze_pump.__version__"}
readme = {file = ["README.md"]}
